+++ packages/in-memory/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Next
+
+- `history.navigate()` takes one argument, an object with an optional `state` property. The object can either contain URL components (`pathname`, `query`, and `hash`) or a URL string under the `url` property.
+- `inMemory`'s `locations` option no longer takes strings. Now, must be a URL object or a URL components object.
+- `createReusable`'s return function's `location` option no longer takes strings. Now, must be a URL object or a URL components object.
+
 ## 2.0.0-beta.11
 
 - Rename `in_memory` to `inMemory`.